feat: Complete Bank contract writing

+++ BankContract/Bank.sol
@@ -6,12 +6,12 @@ contract Bank {
     uint private constant TOP_NUMBERS = 3;
     
     address public owner;
-    mapping (address => uint256) public deposits;
+    mapping (address => uint) public deposits;
     address[3] public topUsers;
-    uint256[3] public topDeposits;
+    uint[3] public topDeposits;
 
-    event Deposit(address indexed user, uint256 amount);
-    event Withdraw(uint256 amount);
+    event Deposit(address indexed user, uint amount);
+    event Withdraw(uint amount);
 
     constructor () {
         owner = msg.sender;
@@ -24,7 +24,7 @@ contract Bank {
         updateTopUsers(msg.sender, deposits[msg.sender]);
     }
 
-    function updateTopUsers(address user, uint256 depositAmount) internal {
+    function updateTopUsers(address user, uint depositAmount) internal {
         for (uint i = 0; i < TOP_NUMBERS; i++) {
             if (depositAmount > topDeposits[i]) {
                 for(uint k = TOP_NUMBERS - 1; k > i; k--) {
@@ -38,13 +38,13 @@ contract Bank {
         }
     }
 
-    function withdraw(uint256 amount) external onlyOwner {
+    function withdraw(uint amount) external onlyOwner {
         require(amount <= address(this).balance, "Insufficient contract balance");
         payable(owner).transfer(amount);
         emit Withdraw(amount);
     }
 
-    function getBalance() external view returns (uint256) {
+    function getBalance() external view returns (uint) {
         return address(this).balance;
     }
